-- keep EVERYBODY happy! BSD has that field signed int wil linux unsigned int
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@14 01689507-d990-df11-9d7d-0017a411d416

+++ sources/thelib/src/protocols/rtp/connectivity/outboundconnectivity.cpp
@@ -294,7 +294,7 @@ bool OutboundConnectivity::InitializePorts(int32_t &dataFd, uint16_t &dataPort,
 	return false;
 }
 #define COMPUTE_BYTES_COUNT(m,c) \
-	for(uint32_t i=0;i<m.msg_iovlen;i++){ \
+	for(uint32_t i=0;i<(uint32_t)m.msg_iovlen;i++){ \
 		c+=m.msg_iov[i].iov_len-12; \
 	}
